requirements.txt
google-cloud-aiplatform>=1.100.0
google-adk>=1.5.0
uvicorn>=0.24.0
fastapi>=0.104.0 
